implemented all snapshot related endpoints in mock (update snapshot/ get delegators / voters) and put new fund endpoint. Fixed typo in vit-ss project and exposed method which combines contributors and voters. Additionally improve rest module organization in mock
