+++ src/vit-servicing-station/vit-servicing-station-lib/src/v0/endpoints/challenges/mod.rs
@@ -4,3 +4,4 @@ mod routes;
 mod schemas;
 
 pub use routes::filter;
+pub use schemas::ChallengeWithProposals;
