+++ src/vit-testing/vitup/src/mode/mock/farm/rest/mod.rs
@@ -2,13 +2,13 @@ use super::ContextError;
 use super::{Context, ContextLock};
 use crate::mode::mock::farm::context::MockId;
 use crate::mode::mock::rest::reject::report_invalid;
+use crate::mode::mock::rest::{load_cert, load_private_key};
 use futures::StreamExt;
+use hyper::service::make_service_fn;
 use jortestkit::web::api_token::TokenError;
 use jortestkit::web::api_token::{APIToken, APITokenManager, API_TOKEN_HEADER};
 use rustls::KeyLogFile;
 use std::convert::Infallible;
-use std::fs::{self, File};
-use std::path::Path;
 use std::sync::Arc;
 use thiserror::Error;
 use tokio::net::TcpListener;
@@ -16,7 +16,6 @@ use tokio_rustls::TlsAcceptor;
 use valgrind::Protocol;
 use vit_servicing_station_lib::v0::result::HandlerResult;
 use warp::http::header::{HeaderMap, HeaderValue};
-use warp::hyper::service::make_service_fn;
 use warp::{reject::Reject, Filter, Rejection, Reply};
 
 impl warp::reject::Reject for ContextError {}
@@ -38,6 +37,8 @@ pub enum Error {
     InvalidCertificate,
     #[error("invalid tls key")]
     InvalidKey,
+    #[error(transparent)]
+    Rest(#[from] crate::mode::mock::rest::Error),
 }
 
 impl Reject for Error {}
@@ -162,34 +163,6 @@ pub async fn start_rest_server(context: ContextLock) -> Result<(), Error> {
             warp::serve(api).bind(address).await;
             Ok(())
         }
-    }
-}
-
-fn load_cert(filename: &Path) -> Result<Vec<rustls::Certificate>, Error> {
-    let certfile = fs::File::open(filename)?;
-    let mut reader = std::io::BufReader::new(certfile);
-
-    match rustls_pemfile::read_one(&mut reader)? {
-        Some(rustls_pemfile::Item::X509Certificate(cert)) => Ok(vec![rustls::Certificate(cert)]),
-        Some(rustls_pemfile::Item::RSAKey(_)) | Some(rustls_pemfile::Item::PKCS8Key(_)) => {
-            // TODO: a more specific error could be useful (ExpectedCertFoundKey?)
-            Err(Error::InvalidCertificate)
-        }
-        // not a pemfile
-        None => Err(Error::InvalidCertificate),
-        Some(_) => Err(Error::InvalidCertificate),
-    }
-}
-
-fn load_private_key(filename: &Path) -> Result<rustls::PrivateKey, Error> {
-    let keyfile = File::open(filename)?;
-    let mut reader = std::io::BufReader::new(keyfile);
-
-    match rustls_pemfile::read_one(&mut reader)? {
-        Some(rustls_pemfile::Item::RSAKey(key)) => Ok(rustls::PrivateKey(key)),
-        Some(rustls_pemfile::Item::PKCS8Key(key)) => Ok(rustls::PrivateKey(key)),
-        None | Some(rustls_pemfile::Item::X509Certificate(_)) => Err(Error::InvalidKey),
-        Some(_) => Err(Error::InvalidCertificate),
     }
 }
