+++ src/vit-testing/vitup/src/mode/mock/rest/control/mod.rs
@@ -1,0 +1,4 @@
+mod handlers;
+mod routes;
+
+pub use routes::control_filter;
