+++ src/vit-testing/vitup/src/mode/mock/rest/node/mod.rs
@@ -1,0 +1,3 @@
+mod handlers;
+
+pub use handlers::*;
